FDA Anaphylaxis Dataset

A 130 by 4 data matrix from the FDA Adverse Event Reporting System (FAERS) database, processed for 2016. It contains drugs and associated adverse events, focusing on events related to anaphylaxis.

data(anaphylaxis)

Format

A data matrix with 130 rows and 4 columns:

drugs

130 drugs represented by rows.

adverse_events

4 adverse events related to anaphylaxis represented by columns.

Details

The 130 rows represent different drugs, while the 4 columns correspond to adverse events associated with anaphylaxis. Data is from the FDA Adverse Event Reporting System (FAERS) database, specifically from 2016 reports.

The dataset helps identify drugs associated with the 4 anaphylaxis-related adverse events, rather than focusing on identifying adverse events associated with a drug class.

Examples

head(anaphylaxis)
#>                    Anaphylactic reaction Anaphylactic shock
#> ABATACEPT                             33                  9
#> ACETAMINOPHEN                         42                 11
#> ADALIMUMAB                           102                 27
#> ALEMTUZUMAB                            1                  0
#> ALENDRONATE SODIUM                     8                  2
#> AMBRISENTAN                            3                  1
#>                    Hypotension_unspecified Urticaria_unspecified
#> ABATACEPT                               50                    55
#> ACETAMINOPHEN                          152                   116
#> ADALIMUMAB                             285                  1662
#> ALEMTUZUMAB                             43                    87
#> ALENDRONATE SODIUM                     161                    25
#> AMBRISENTAN                            501                    17