FDA dataset for beta blockers with 500 adverse events

A 501 by 9 data matrix of a contingency table processed from the FDA Adverse Event Reporting System (FAERS) database. This dataset covers a specific period from Q1 2021 to Q4 2023.

Usage

betablocker500

Format

A data matrix with 501 rows and 9 columns.

Details

A 501 by 9 data matrix of a contingency table from the FDA Adverse Event Reporting System (FAERS) database, covering the period from Q1 2021 to Q4 2023.

The 500 rows correspond to the AEs with the highest overall frequency (row marginals) reported during the period, and 1 row for Other AEs. The reported AEs - "Off label use" and "Drug ineffective" have been excluded.

The dataset includes the following 9 columns: Acebutolol, Atenolol, Bisoprolol, Carvedilol, Metoprolol, Nadolol, Propranolol, Timolol, and Other.

The marginal totals for each column are as follows: Acebutolol: 62,164, Atenolol: 36,619, Bisoprolol: 134,297, Carvedilol: 35,922, Metoprolol: 88,387, Nadolol: 11,191, Propranolol: 56,444, Timolol: 16,077, Other: 76,926,859.

Also refer to the supplementary material of:

Ding, Y., Markatou, M., & Ball, R. (2020). An evaluation of statistical approaches to postmarketing surveillance. Statistics in Medicine, 39(7), 845-874

for the data generation process. The quarterly files can be found in https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html.

Examples

data(betablocker500)
head(betablocker500)
#>                      Acebutolol Atenolol Bisoprolol Carvedilol Metoprolol
#> Pain                       3582      455        977        110        457
#> Fatigue                       0      503       1830        621       1191
#> Nausea                       13      994       1313        122        816
#> Arthralgia                   10      502        892        131        617
#> Condition aggravated          1      474        409         90        365
#> Dyspnoea                   1122      305       2368        549       1702
#>                      Nadolol Propranolol Timolol   Other
#> Pain                     307         166     348 1074063
#> Fatigue                   34         459      79 1015237
#> Nausea                   349        1023      51  757910
#> Arthralgia                17         188      42  688797
#> Condition aggravated       7         359     393  679423
#> Dyspnoea                  76         411     385  672028