Generate simulated contingency tables with option of incorporating adverse event correlation within clusters.

Generate simulated contingency tables with option of incorporating adverse event correlation within clusters.

generate_contin_table_with_clustered_AE(
  contin_table,
  n_rep = 1,
  AE_idx,
  rho = 0.5,
  signal_mat
)

Arguments

contin_table

A data matrix of an \(I\) x \(J\) contingency table with row (adverse event) and column (drug) names, of which the row and column marginals are used to generate the simulated data. Please first check the input contingency table using the function check_and_fix_contin_table().

n_rep

Number of contingency tables to be generated.

AE_idx

A data frame with two variables idx and AE, where idx indicates the cluster index (can be either a name or a number), and AE lists the adverse event names. See the AE_idx for statin49 data as an example.

rho

A numeric value indicating the correlation of the AEs within each cluster. Default is 0.5.

signal_mat

A data matrix of the same dimension as the contingency table with entries representing the signal strength. The values should be greater or equal to 1, where 1 indicates no signal, and values greater than 1 indicate signal.

Value

A list of n_rep simulated contingency tables.

Examples

# using statin49 as an example
data(statin49)
data(AE_idx)

# Prepare a matrix of signal strength with the same dimension as
# statin49, where 1 indicates no signal and values > 1 indicate
# signal
lambda_matrix <- matrix(1, nrow = nrow(statin49), ncol = ncol(statin49))

# Assign the cell (45,1) with signal strength 4
lambda_matrix[45, 1] <- 4

# Generate 5 simulated tables
set.seed(123)
simulated_tables <- generate_contin_table_with_clustered_AE(
  contin_table = statin49,
  n_rep = 5,
  AE_idx = AE_idx,
  rho = 0.5,
  signal_mat = lambda_matrix
)