FDA dataset for sedatives with 1000 adverse events

A 1001 by 11 data matrix of a contingency table processed from the FDA Adverse Event Reporting System (FAERS) database. This dataset covers a specific period from Q1 2021 to Q4 2023.

Usage

sedative1000

Format

A data matrix with 1001 rows and 11 columns.

Details

A 1001 by 11 data matrix of a contingency table from the FDA Adverse Event Reporting System (FAERS) database, covering a specified period from Q1 2021 to Q4 2023.

The 1000 rows correspond to the AEs with the highest overall frequency (row marginals) reported during the period and 1 row for Other AEs. The reported AEs - "Off label use" and "Drug ineffective" have been excluded.

The dataset includes the following 10 columns: Clonazepam, Dexmedetomidine, Diazepam, Diphenhydramine, Doxepin, Lorazepam, Midazolam, Mirtazapine, Nitrazepam, Temazepam, and an Other column.

The marginal totals for each column are as follows: Clonazepam: 110,453, Dexmedetomidine: 4,262, Diazepam: 74,859 Diphenhydramine: 134,65, Doxepin: 11,795, Lorazepam: 101,969 Midazolam: 26,264, Mirtazapine: 54,273, Nitrazepam: 3,473, Temazepam: 20,523, Other: 77,487,518

Also refer to supplementary material of: Ding, Y., Markatou, M., & Ball, R. (2020). An evaluation of statistical approaches to postmarketing surveillance. Statistics in Medicine, 39(7), 845-874

for the data generation process. The quarterly files can be found in https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html.

Examples

data(sedative1000)
head(sedative1000)
#>                      Clonazepam Dexmedetomidine Diazepam Diphenhydramine
#> Pain                        781              16      617            1963
#> Fatigue                    1014               7      481            2074
#> Nausea                      593               5      630            1532
#> Arthralgia                  454               0      229            1190
#> Condition aggravated        462              27      601             708
#> Dyspnoea                   1412               7      532             611
#>                      Doxepin Lorazepam Midazolam Mirtazapine Nitrazepam
#> Pain                     194       427       195         212         11
#> Fatigue                   89      1729        37         526         37
#> Nausea                    38       842       228         453         13
#> Arthralgia                62       336        38         375          7
#> Condition aggravated      35       411       181         250         14
#> Dyspnoea                  40       824       115         341         27
#>                      Temazepam   Other
#> Pain                       397 1075677
#> Fatigue                    363 1013681
#> Nausea                      42  758316
#> Arthralgia                  27  688506
#> Condition aggravated        70  678843
#> Dyspnoea                    16  675191