FDA statin dataset with 49 adverse events

A 50 by 7 data matrix of a contingency table processed from the FDA Adverse Event Reporting System (FAERS) database from Q3 2014 to Q4 2020.

Usage

statin49

Format

A data matrix with 50 rows and 7 columns.

Details

A 50 by 7 data matrix of a contingency table from the FDA Adverse Event Reporting System (FAERS) database, covering Q3 2014 to Q4 2020.

The 49 rows represent important adverse events associated with statins, with the final row aggregating the remaining 5,990 events. The 49 AEs are classified into three clusters (see statin49_AE_idx for cluster indices):

1) AEs associated with muscle injury signs and symptoms, 2) AEs associated with muscle injury lab tests, and 3) AEs associated with kidney injury and its diagnosis and treatment.

The 7 columns include six statin medications and an "other drugs" column for occurrences of the 6,039 events. Marginal totals for each drug: 197,390 for Atorvastatin, 5,742 for Fluvastatin, 3,230 for Lovastatin, 22,486 for Pravastatin, 122,450 for Rosuvastatin, 85,445 for Simvastatin, and 63,539,867 for Other drugs.

Examples

data(statin49)
head(statin49)
#>                    Atorvastatin Fluvastatin Lovastatin Pravastatin Rosuvastatin
#> Rhabdomyolysis             2041          52         44         163          936
#> Muscle Disorder             291           2          7          21          191
#> Muscle Fatigue               85           0          2          16           30
#> Muscle Haemorrhage           24           0          0           5           13
#> Muscle Necrosis              68           2          0           1           10
#> Muscle Rupture              181          25          0          61           36
#>                    Simvastatin Other
#> Rhabdomyolysis            1376 31707
#> Muscle Disorder             87  7329
#> Muscle Fatigue              39  4257
#> Muscle Haemorrhage           4  3806
#> Muscle Necrosis             20   662
#> Muscle Rupture             120  3219