Cluster index of the FDA statin dataset with 49 adverse events

A 50 by 2 data frame showing the cluster index of the statin49 dataset's adverse events.

Usage

statin49_AE_idx

Format

A data frame with 50 rows and 2 columns: idx and AE.

Details

A data frame with 50 rows and 2 columns: idx and AE. AE lists the adverse event names in the statin49 dataset, while idx lists the cluster index of each adverse event.

The 49 AEs are classified into three clusters: 1) AEs associated with signs and symptoms of muscle injury, 2) AEs associated with laboratory tests for muscle injury, and 3) AEs associated with kidney injury and its laboratory diagnosis and treatment.

Examples

data(statin49_AE_idx)
head(statin49_AE_idx)
#>   idx                 AE
#> 1   1     Rhabdomyolysis
#> 2   1    Muscle Disorder
#> 3   1     Muscle Fatigue
#> 4   1 Muscle Haemorrhage
#> 5   1    Muscle Necrosis
#> 6   1     Muscle Rupture